Refactor userSchema.statics.findByCredentials to return the User.
I created some really weird patterns when I first made this app. I
was always returning these dbResponse objects that had users and
statusCodes or statusMessages attached. This was strange. I'm fixing
it now to return sane things like the User, or throw a ResponseError
if there is an error.